Changed bits type.

+++ bitcopy.c
@@ -31,17 +31,17 @@
 #include <stdlib.h>
 
 void bitcopy(void *dst, const void *src,
-	     const unsigned long bits[], const size_t length);
+	     const size_t bits[], const size_t length);
 
 void bitcopy(void *dst, const void *src,
-	     const unsigned long bits[], const size_t length)
+	     const size_t bits[], const size_t length)
 {
   char *d = dst;
   const char *s = src;
 
   for(size_t i = 0; i < length; i++)
     {
-      unsigned long bit = bits[i] % 8;
+      size_t bit = bits[i] % 8;
       size_t idx = bits[i] >> 3; // Divide by eight.
 
       if(((1 << bit) & s[idx])) // Is the bit set?
@@ -57,37 +57,37 @@ int main(void)
   char b[] = "Please see the doctor!";
   int32_t x = 101;
   int32_t y = 202;
-  unsigned long bits_ab[] = {8 * 4,
-			     8 * 4 + 1,
-			     8 * 4 + 2,
-			     8 * 4 + 3,
-			     8 * 4 + 4,
-			     8 * 4 + 5,
-			     8 * 4 + 6,
-			     8 * 4 + 7,
-			     8 * 20,
-			     8 * 20 + 1,
-			     8 * 20 + 2,
-			     8 * 20 + 3,
-			     8 * 20 + 4,
-			     8 * 20 + 5,
-			     8 * 20 + 6,
-			     8 * 20 + 7};
-  unsigned long bits_xy[] = {0, 1, 2, 3, 4, 5, 6, 7,
-			     8 * 3,
-			     8 * 3 + 1,
-			     8 * 3 + 2,
-			     8 * 3 + 3,
-			     8 * 3 + 4,
-			     8 * 3 + 5,
-			     8 * 3 + 6,
-			     8 * 3 + 7};
+  size_t bits_ab[] = {8 * 4,
+		      8 * 4 + 1,
+		      8 * 4 + 2,
+		      8 * 4 + 3,
+		      8 * 4 + 4,
+		      8 * 4 + 5,
+		      8 * 4 + 6,
+		      8 * 4 + 7,
+		      8 * 20,
+		      8 * 20 + 1,
+		      8 * 20 + 2,
+		      8 * 20 + 3,
+		      8 * 20 + 4,
+		      8 * 20 + 5,
+		      8 * 20 + 6,
+		      8 * 20 + 7};
+  size_t bits_xy[] = {0, 1, 2, 3, 4, 5, 6, 7,
+	 	      8 * 3,
+		      8 * 3 + 1,
+		      8 * 3 + 2,
+		      8 * 3 + 3,
+		      8 * 3 + 4,
+		      8 * 3 + 5,
+		      8 * 3 + 6,
+		      8 * 3 + 7};
 
   printf("a = %s, b = %s.\n", a, b);
-  bitcopy(a, b, bits_ab, sizeof(bits_ab) / sizeof(unsigned long));
+  bitcopy(a, b, bits_ab, sizeof(bits_ab) / sizeof(size_t));
   printf("a = %s, b = %s.\n", a, b);
   printf("x = %d, y = %d.\n", x, y);
-  bitcopy(&x, &y, bits_xy, sizeof(bits_xy) / sizeof(unsigned long));
+  bitcopy(&x, &y, bits_xy, sizeof(bits_xy) / sizeof(size_t));
   printf("x = %d, y = %d.\n", x, y);
   return EXIT_SUCCESS;
 }
